Fixed typing bug that caused fet % to be mis displayed.

+++ src/firmware/draining_battery_ui.c
@@ -129,8 +129,8 @@ static void render_line3(struct SharedState* ss, uint8_t active) {
   text_strLen(text, line, 4);
   text->options = 0x00;
 
-  uint8_t fet_centi_percent =
-    active ? (uint8_t)(((uint32_t)(ss->vgs_level) * 100) / 655) :
+  uint16_t fet_centi_percent =
+    active ? (uint16_t)(((uint32_t)(ss->vgs_level) * 100) / 655) :
     ss->max_values.fet_percent * 100;
   if (active && ss->is_sampling_voltage) {
     fet_centi_percent = 0;
